Settled on a stable scheme for extreme pipelining.  Sample req-ack can be 0-delay, but Update req-ack will have unit delay.  The reenables to sample-start will have unit delay, and the reenables to update-start will have 0-delay.
Also, fixed a bug in the memory subsystem.  Since store->load dependencies
are maintained, a simultaneous load and store reaching the memory
subsystem should be treated as if the store has priority. The best
solution may be to introduce a single cycle latency from load
sample-complete to successive store sample-start?