Looks like with a little bit of help, we can get a processor core running at 1 IPC by describing the core in Aa.
The register-file and memory-subsystem
will have to be coded manually in RTL so that low latency, high
throughput performance accesses can be supported.  Some
care needs to be taken to ensure buffering in the processor AA
code.

The Rfile needs 3 (actually 2) read ports and 1 write port.
The Memory subsystem will be separated into D and I streams
with one R/W port on each.